fix: brought out the logic for printing an object in ObjectPrinter class
