+++ ObjectPrinting/PrintingConfigs/Extensions/ObjectExtensions.cs
@@ -6,12 +6,12 @@ public static class ObjectExtensions
 {
     public static string PrintToString<T>(this T obj)
     {
-        return ObjectPrinter.For<T>().PrintToString(obj);
+        return ObjectPrinter.For<T>().CreatePrinter().PrintToString(obj);
     }
     
     public static string PrintToString<T>(this T obj, Func<PrintingConfig<T>, PrintingConfig<T>> config)
     {
-        return config(ObjectPrinter.For<T>()).PrintToString(obj);
+        return config(ObjectPrinter.For<T>()).CreatePrinter().PrintToString(obj);
     }
 }
